+++ packages/cli/src/commands/asset/index.ts
@@ -1,9 +1,12 @@
-import { commandDirOptions } from '@appsemble/node-utils';
 import { Argv } from 'yargs';
+
+import * as create from './create';
+
+export { noop as handler } from '@appsemble/utils';
 
 export const command = 'asset';
 export const description = 'Commands related to app assets.';
 
 export function builder(yargs: Argv): Argv {
-  return yargs.commandDir(...commandDirOptions(__filename));
+  return yargs.command(create).demandCommand(1);
 }