+++ packages/cli/src/commands/config/index.ts
@@ -1,9 +1,13 @@
-import { commandDirOptions } from '@appsemble/node-utils';
 import { Argv } from 'yargs';
+
+import * as get from './get';
+import * as set from './set';
+
+export { noop as handler } from '@appsemble/utils';
 
 export const command = 'config';
 export const description = 'Commands related to local configurations.';
 
 export function builder(yargs: Argv): Argv {
-  return yargs.commandDir(...commandDirOptions(__filename));
+  return yargs.command(get).command(set).demandCommand(1);
 }