+++ packages/server/src/controllers/appOAuth2Secrets.test.ts
@@ -367,6 +367,82 @@ describe('updateAppOAuth2Secret', () => {
   });
 });
 
+describe('deleteAppOAuth2Secret', () => {
+  it('should delete an OAuth2 secret', async () => {
+    const secret = await AppOAuth2Secret.create({
+      AppId: app.id,
+      authorizationUrl: 'https://example.com/oauth/authorize',
+      clientId: 'example_client_id',
+      clientSecret: 'example_client_secret',
+      icon: 'example',
+      name: 'Example',
+      scope: 'email openid profile',
+      tokenUrl: 'https://example.com/oauth/token',
+      userInfoUrl: 'https://example.com/oauth/userinfo',
+    });
+    authorizeStudio();
+    const response = await request.delete(`/api/apps/${app.id}/secrets/oauth2/${secret.id}`);
+    expect(response).toMatchInlineSnapshot('HTTP/1.1 204 No Content');
+  });
+
+  it('should handle if the app id is invalid', async () => {
+    authorizeStudio();
+    const response = await request.delete('/api/apps/123/secrets/oauth2/1');
+    expect(response).toMatchInlineSnapshot(`
+      HTTP/1.1 404 Not Found
+      Content-Type: application/json; charset=utf-8
+
+      {
+        "error": "Not Found",
+        "message": "App not found",
+        "statusCode": 404,
+      }
+    `);
+  });
+
+  it('should handle if the secret id is invalid', async () => {
+    authorizeStudio();
+    const response = await request.delete(`/api/apps/${app.id}/secrets/oauth2/1`);
+    expect(response).toMatchInlineSnapshot(`
+      HTTP/1.1 404 Not Found
+      Content-Type: application/json; charset=utf-8
+
+      {
+        "error": "Not Found",
+        "message": "OAuth2 secret not found",
+        "statusCode": 404,
+      }
+    `);
+  });
+
+  it('should require the user to have correct permissions', async () => {
+    const secret = await AppOAuth2Secret.create({
+      AppId: app.id,
+      authorizationUrl: 'https://example.com/oauth/authorize',
+      clientId: 'example_client_id',
+      clientSecret: 'example_client_secret',
+      icon: 'example',
+      name: 'Example',
+      scope: 'email openid profile',
+      tokenUrl: 'https://example.com/oauth/token',
+      userInfoUrl: 'https://example.com/oauth/userinfo',
+    });
+    await member.update({ role: 'Member' });
+    authorizeStudio();
+    const response = await request.delete(`/api/apps/${app.id}/secrets/oauth2/${secret.id}`);
+    expect(response).toMatchInlineSnapshot(`
+      HTTP/1.1 403 Forbidden
+      Content-Type: application/json; charset=utf-8
+
+      {
+        "error": "Forbidden",
+        "message": "User does not have sufficient permissions.",
+        "statusCode": 403,
+      }
+    `);
+  });
+});
+
 describe('verifyAppOAuth2SecretCode', () => {
   let secret: AppOAuth2Secret;