+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## \[[0.20.8](https://gitlab.com/appsemble/appsemble/-/releases/0.20.8)] - 2022-06-12
+
+### Added
+
+- Block(`action-button`): Add `title` property.
+- Block(`button-list`): Add `title` property.
+- Block(`form`): Add `color` and `size` properties to `boolean` fields.
+- Block(`form`): Add support for displaying boolean/checkbox fields as switches.
+- Block(`table`): Add `disabled` property for button columns.
+- Block(`table`): Add `title` property to button fields.
+- App: Add `view` property to `resource.query` and `resource.get` actions.
+- Server: Add support for resource views.
+- Utils: Add `object.omit` remapper.
+- Utils: Add `random.float` `random.integer` `random.string` remapper.
+
+### Fixed
+
+- Block(`form`): Fix bug where `icon` would be ignored for `date` and `date-time` fields.
+- Block(`table`): Move `alignment` property specific to dropdown fields to the `dropdown` field
+  property. This was added to `DropdownOption` instead of `DropdownField` by accident.
+- App: Fix bug where register form was displayed twice.
+
 ## \[[0.20.7](https://gitlab.com/appsemble/appsemble/-/releases/0.20.7)] - 2022-05-25
 
 ### Added